requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
sqlmodel==0.0.37
pwdlib==0.3.0
PyJWT==2.11.0
pydantic[email]==2.12.5
python-multipart==0.0.22
